kv: Move TestEngineBuilder to its own module
